chore(storybook):  改用 vite 框架
(import.meta as any).env 代替 process.env
ts 的类型声明需要特殊处理
ts 的 module 配置改用 ESNext
移除不需要的依赖项 @storybook/react-webpack5 和 react-docgen-typescript-loader
