+++ package.json
@@ -49,7 +49,7 @@
     "@storybook/addon-actions": "^7.5.3",
     "@storybook/addon-essentials": "^7.5.3",
     "@storybook/react": "^7.5.3",
-    "@storybook/react-webpack5": "^7.5.3",
+    "@storybook/react-vite": "^7.5.3",
     "@storybook/storybook-deployer": "^2.8.16",
     "@testing-library/react": "^12.1.5",
     "@testing-library/react-hooks": "^8.0.1",
@@ -72,7 +72,6 @@
     "jest-environment-jsdom": "^29.7.0",
     "lint-staged": "^13.0.3",
     "prettier": "^2.0.5",
-    "react-docgen-typescript-loader": "^3.7.2",
     "react-intersection-observer": "^9.5.2",
     "rollup": "^2.40.0",
     "rollup-plugin-filesize": "^10.0.0",
