+++ src/components/AMapAPIContainer/index.ts
@@ -1,6 +1,3 @@
-export { AMapAPIContextValue, default as AMapAPIContext } from './AMapAPIContext';
-export {
-  CreateAMapApiContainerOptions,
-  AMapAPIHocProps,
-  createAMapAPIContainer,
-} from './createAMapAPIContainer';
+export type { AMapAPIContextValue, AMapAPIHocProps } from './interface';
+export { default as AMapAPIContext } from './AMapAPIContext';
+export { createAMapAPIContainer } from './createAMapAPIContainer';
