+++ src/components/AMapControlBar/AMapControlBar.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react/default-props-match-prop-types */
 import type { FC } from 'react';
 import {
   useCallback,
@@ -10,20 +11,13 @@ import useAMapControlBinder from '../../hooks/useAMapControlBinder';
 import useAMapEventBinder from '../../hooks/useAMapEventBinder';
 import useVisible from '../../hooks/useVisible';
 
+import type { AMapControlBarProps } from './interface';
 import useControlButtonUpdater from './useControlButtonUpdater';
 
 /**
  * Origin API see:
  * https://lbs.amap.com/api/jsapi-v2/documentation#ControlBar
  */
-
-export type AMapControlBarProps = Pick<AMap.ControlBarConfig, 'offset'> & {
-  position?: AMap.ControlBarConfig['position'];
-  showControlButton?: boolean;
-  visible?: boolean;
-  onShow?: (event: any) => void;
-  onHide?: (event: any) => void;
-};
 
 const defaultProps = {
   position: 'LT' as AMapControlBarProps['position'],
