+++ src/components/AMapMap/index.ts
@@ -1,3 +1,4 @@
-export * from './interface';
+export type { AMapMapContextValue } from './interface';
+export type { AMapMapProps } from './interface';
 export { default as AMapMapContext } from './AMapMapContext';
 export { default, default as AMapMap } from './AMapMap';
