+++ src/components/AMapToolBar/interface.ts
@@ -1,0 +1,6 @@
+export type AMapToolBarProps = Pick<AMap.ToolBarConfig, 'offset'> & {
+  position?: AMap.ControlBarConfig['position'];
+  visible?: boolean;
+  onShow?: (event?: any) => void;
+  onHide?: (event?: any) => void;
+};
